+++ src/NFS_Protocols/nfs41_op_read.c
@@ -235,7 +235,7 @@ int nfs41_op_read(struct nfs_argop4 *op, compound_data_t * data, struct nfs_reso
   offset = arg_READ4.offset;
   size = arg_READ4.count;
 
-  LogFullDebug(COMPONENT_NFS_V4, "   NFS4_OP_READ: offset = %llu  length = %llu", offset, size);
+  LogFullDebug(COMPONENT_NFS_V4, "   NFS4_OP_READ: offset = %llu  length = %llu", (unsigned long long)offset, size);
 
   if((data->pexport->options & EXPORT_OPTION_MAXOFFSETREAD) ==
      EXPORT_OPTION_MAXOFFSETREAD)
@@ -330,7 +330,7 @@ int nfs41_op_read(struct nfs_argop4 *op, compound_data_t * data, struct nfs_reso
   res_READ4.READ4res_u.resok4.data.data_len = read_size;
   res_READ4.READ4res_u.resok4.data.data_val = bufferdata;
 
-  LogFullDebug(COMPONENT_NFS_V4,"   NFS4_OP_READ: offset = %llu  read length = %llu eof=%u", offset, read_size,
+  LogFullDebug(COMPONENT_NFS_V4,"   NFS4_OP_READ: offset = %llu  read length = %llu eof=%u", (unsigned long long)offset, read_size,
          eof_met);
 
   /* Is EOF met or not ? */
